+++ mirage/config.ts
@@ -2,7 +2,6 @@ import { Server } from 'ember-cli-mirage';
 import config from 'ember-get-config';
 
 import { getCitation } from './views/citation';
-import { getProviderTaxonomies } from './views/collection-provider-taxonomies';
 import { searchCollections } from './views/collection-search';
 import { reportDelete } from './views/comment';
 import { createDeveloperApp, resetClientSecret } from './views/developer-app';
@@ -150,7 +149,6 @@ export default function(this: Server) {
     osfNestedResource(this, 'collection', 'collectedMetadata', {
         path: 'collections/:parentID/collected_metadata/',
     });
-    this.get('/providers/collections/:parentID/taxonomies', getProviderTaxonomies);
     this.post('/search/collections/', searchCollections);
 
     // Waterbutler namespace